requirements.txt
aiohttp==3.9.3
aioice==0.9.0
aiortc==1.8.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
cryptography==42.0.7
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.1
fastjsonschema==2.19.1
filelock==3.13.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-crc32c==1.5.0
google-search-results==2.4.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kubernetes==29.0.0
langchain==0.1.15
langchain-chroma==0.1.1
langchain-community==0.0.32
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.0.1
langchain-upstage==0.1.5
langchainhub==0.1.15
langserve==0.0.51
langsmith==0.1.43
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.30.3
openai-whisper==20231117
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.10.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.27.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
streamlit-audiorec==0.1.3
streamlit-webrtc==0.47.7
sympy==1.12
tenacity==8.2.3
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
yt-dlp==2024.5.27
zipp==3.18.1
pysqlite3-binary